renamer suffix på alle behandlingsfaner fra Side til Fane. Fjerner også default export på alle fanene
